# api-admin/src/main/resources/api/openapi.yml
openapi: 3.0.3
info:
  title: Juniper Admin API
  description: API for portal and study management
  version: 0.0.1
paths:
...
  /api/portals/v1/{portalShortcode}/studies/{studyShortcode}/env/{envName}/requestKits:
    post:
      summary: Request sample kits for a batch of enrollees
      tags: [ kit ]
      operationId: requestKits
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
        - { name: studyShortcode, in: path, required: true, schema: { type: string } }
        - { name: envName, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        description: kit request list creation dto
        content: { application/json: { schema: { type: object } } }
      responses:
        '200':
          description: Kit requests
          content: { application/json: { schema: { type: object } } }
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/portals/v1/{portalShortcode}/studies/{studyShortcode}/env/{envName}/enrollees/{enrolleeShortcode}/kitRequests:
    get:
      summary: List kit requests for an enrollee
      tags: [ kit ]
      operationId: getKitRequests
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
        - { name: studyShortcode, in: path, required: true, schema: { type: string } }
        - { name: envName, in: path, required: true, schema: { type: string } }
        - { name: enrolleeShortcode, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Kit requests
          content: { application/json: { schema: { type: object } } }
        '500':
          $ref: '#/components/responses/ServerError'
  /api/portals/v1/{portalShortcode}/studies/{studyShortcode}/env/{envName}/enrollees/{enrolleeShortcode}/requestKit:
    post:
      summary: Request a sample kit for an enrollee
      tags: [ kit ]
      operationId: requestKit
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
        - { name: studyShortcode, in: path, required: true, schema: { type: string } }
        - { name: envName, in: path, required: true, schema: { type: string } }
        - { name: enrolleeShortcode, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        description: kit request creation dto
        content: { application/json: { schema: { type: object } } }
      responses:
        '200':
          description: Kit request
          content: { application/json: { schema: { type: object } } }
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/address/v1/validate:
    put:
      summary: Validate a given address
      tags: [ addressValidation ]
      operationId: validate
      requestBody:
        description: mailing address
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        '200':
          description: validation results
          content: { application/json: { schema: { type: object } } }
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/internal/v1/populate/baseSeed:
    post:
      summary: Populates
      tags: [ populate ]
      operationId: populateBaseSeed
      responses:
        '201':
          description: stats object
          content:
            application/json:
              schema:
                type: object
        '500':
          $ref: '#/components/responses/ServerError'
  /api/internal/v1/populate/adminConfig:
    post:
      summary: Populates admin config
      tags: [ populate ]
      operationId: populateAdminConfig
      parameters:
        - { name: overwrite, in: query, required: false, schema: { type: boolean } }
      responses:
        '201':
          description: summary stats for the operation
          content:
            application/json:
              schema:
                type: object
        '500':
          $ref: '#/components/responses/ServerError'
  /api/internal/v1/populate/portal:
    post:
      summary: Populates a portal from a flat file that is already on the server
      tags: [ populate ]
      operationId: populatePortal
      parameters:
        - { name: filePathName, in: query, required: true, schema: { type: string } }
        - { name: overwrite, in: query, required: false, schema: { type: boolean } }
        - { name: shortcodeOverride, in: query, required: false, schema: { type: string } }
      responses:
        '201':
          description: portal object
          content:
            application/json:
              schema:
                type: object
        '500':
          $ref: '#/components/responses/ServerError'
  /api/internal/v1/populate/portal/upload:
    post:
      summary: upload a zip file of a portal (typically generated from the extract endpoint)
      tags: [ populate ]
      operationId: uploadPortal
      parameters:
        - { name: overwrite, in: query, required: false, schema: { type: boolean } }
        - { name: shortcodeOverride, in: query, required: false, schema: { type: string } }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: portal object (abbreviated)
          content: { application/json: { schema: { type: object } } }
        '500':
          $ref: '#/components/responses/ServerError'
  /api/internal/v1/populate/portal/{portalShortcode}/extract:
    get:
      summary: Extract portal configs as a zip file
      tags: [ populate ]
      operationId: extractPortal
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: zip file of portal configs
          content: { text/plain: { schema: { type: string, format: binary } } }
        '500':
          $ref: '#/components/responses/ServerError'
  /api/internal/v1/populate/siteContent/{portalShortcode}:
    post:
      summary: Populates a siteContent to a given portal
      tags: [ populate ]
      operationId: populateSiteContent
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
        - { name: filePathName, in: query, required: true, schema: { type: string } }
        - { name: overwrite, in: query, required: false, schema: { type: boolean } }
      responses:
        '201':
          description: siteContent object
          content:
            application/json:
              schema:
                type: object
        '500':
          $ref: '#/components/responses/ServerError'
  /api/internal/v1/populate/survey/{portalShortcode}:
    post:
      summary: Populates a survey to a given portal
      tags: [ populate ]
      operationId: populateSurvey
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
        - { name: filePathName, in: query, required: true, schema: { type: string } }
        - { name: overwrite, in: query, required: false, schema: { type: boolean } }
      responses:
        '201':
          description: survey object
          content:
            application/json:
              schema:
                type: object
        '500':
          $ref: '#/components/responses/ServerError'
  /api/internal/v1/populate/enrollee/{portalShortcode}/env/{envName}/study/{studyShortcode}:
    post:
      summary: Populates an Enrollee to a given portal/study/env.
      tags: [ populate ]
      operationId: populateEnrollee
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
        - { name: envName, in: path, required: true, schema: { type: string } }
        - { name: studyShortcode, in: path, required: true, schema: { type: string } }
        - { name: filePathName, in: query, required: true, schema: { type: string } }
        - { name: overwrite, in: query, required: false, schema: { type: boolean } }
      responses:
        '201':
          description: enrollee object
          content:
            application/json:
              schema:
                type: object
        '500':
          $ref: '#/components/responses/ServerError'
  /api/internal/v1/populate/enrollee/{portalShortcode}/env/{envName}/study/{studyShortcode}/bulk:
    post:
      summary: Bulk populates enrollees in a given portal/study/env.
      tags: [ populate ]
      operationId: bulkPopulateEnrollees
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
        - { name: envName, in: path, required: true, schema: { type: string } }
        - { name: studyShortcode, in: path, required: true, schema: { type: string } }
        - { name: numEnrollees, in: query, required: true, schema: { type: integer } }
      responses:
        '204':
          description: No Content
        '500':
          $ref: '#/components/responses/ServerError'
  /api/internal/v1/populate/command/{command}:
    post:
      summary: Runs a specified populate command
      tags: [ populate ]
      operationId: populateCommand
      parameters:
        - { name: command, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content: { application/json: { schema: { type: object } } }
      responses:
        '201':
          description: command result object
          content:
            application/json:
              schema:
                type: object
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/portals/v1/{portalShortcode}/studies/{studyShortcode}/env/{envName}/kits:
    get:
      summary: Gets a list of kits for a study
      tags: [ kit ]
      operationId: kitsByStudyEnvironment
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
        - { name: studyShortcode, in: path, required: true, schema: { type: string } }
        - { name: envName, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: kit list
          content: { application/json: { schema: { type: object } } }
        '500':
          $ref: '#/components/responses/ServerError'
  /api/portals/v1/{portalShortcode}/studies/{studyShortcode}/env/{envName}/kits/refreshKitStatuses:
    post:
      summary: Fetch kit statuses from Pepper and cache in the Juniper database
      tags: [ kit ]
      operationId: refreshKitStatuses
      parameters:
        - { name: portalShortcode, in: path, required: true, schema: { type: string } }
        - { name: studyShortcode, in: path, required: true, schema: { type: string } }
        - { name: envName, in: path, required: true, schema: { type: string } }
      responses:
        '204':
          description: Success
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
...
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReport'
  schemas:
...
    ErrorReport:
      type: object
      required: [ message, statusCode ]
      properties:
        message:
          type: string
        errorClass:
          type: string
        statusCode:
          type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: [ ]